break up the loader module - move prep_data_dirs and remove_project_input_files to utils, and add TODO_AFTER_DATASET_RELEASE to the methods for later removal
