+++ api/scpca_portal/utils/helpers.py
@@ -2,15 +2,44 @@
 
 import inspect
 import math
+import shutil
 from collections import namedtuple
 from datetime import datetime
 from pathlib import Path
 from typing import Any, Callable, Dict, Generator, Iterable, List, Set, Tuple
 
+from django.conf import settings
+
 from scpca_portal import common
 from scpca_portal.config.logging import get_and_configure_logger
 
 logger = get_and_configure_logger(__name__)
+
+
+def prep_data_dirs(wipe_input_dir: bool = False, wipe_output_dir: bool = True) -> None:
+    """
+    Create the input and output data dirs, if they do not yet exist.
+    Allow for options to be passed to wipe these dirs if they do exist.
+        - wipe_input_dir defaults to False because we typically want to keep input data files
+        between testing rounds to speed up our tests.
+        - wipe_output_dir defaults to True because we typically don't want to keep around
+        computed files after execution.
+    The options are given to the caller for to customize behavior for different use cases.
+    """
+    # Prepare data input directory.
+    if wipe_input_dir:
+        shutil.rmtree(settings.INPUT_DATA_PATH, ignore_errors=True)
+    settings.INPUT_DATA_PATH.mkdir(exist_ok=True, parents=True)
+
+    # Prepare data output directory.
+    if wipe_output_dir:
+        shutil.rmtree(settings.OUTPUT_DATA_PATH, ignore_errors=True)
+    settings.OUTPUT_DATA_PATH.mkdir(exist_ok=True, parents=True)
+
+
+def remove_project_input_files(project_id: str) -> None:
+    """Remove the input files located at the project_id's input directory."""
+    shutil.rmtree(settings.INPUT_DATA_PATH / project_id, ignore_errors=True)
 
 
 def boolean_from_string(value: str) -> bool:
